requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
packaging==23.2
pep8==1.7.1
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyOpenSSL==23.3.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
zope.interface==6.1